+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "portfolio"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 
 [lib]
 crate-type = ["cdylib", "rlib"]
