+++ src/components/mod.rs
@@ -1,5 +1,5 @@
-pub mod language_icon;
+pub mod icon;
 pub mod repo;
 
-pub use language_icon::LanguageIcon;
+pub use icon::LanguageIcon;
 pub use repo::{Repo, RepoLoading};
