+++ src/main.rs
@@ -1,7 +1,7 @@
 use axum::Router;
-use islands::app::{shell, App};
 use leptos::prelude::*;
-use leptos_axum::{generate_route_list, LeptosRoutes};
+use leptos_axum::{LeptosRoutes, generate_route_list};
+use portfolio::app::{App, shell};
 
 #[tokio::main]
 async fn main() {
